+++ Source/CesiumRuntime/Private/CesiumGltfComponent.cpp
@@ -1165,13 +1165,16 @@ static void loadMesh(
   const Mesh& mesh = *options.pMesh;
 
   result = LoadMeshResult();
-
+  result->primitiveResults.reserve(mesh.primitives.size());
   for (const CesiumGltf::MeshPrimitive& primitive : mesh.primitives) {
     CreatePrimitiveOptions primitiveOptions = {&options, &primitive};
-    loadPrimitive(
-        result->primitiveResults.emplace_back(),
-        transform,
-        primitiveOptions);
+    auto& primitiveResult = result->primitiveResults.emplace_back();
+    loadPrimitive(primitiveResult, transform, primitiveOptions);
+
+    // if it doesn't have render data, then it can't be loaded
+    if (!primitiveResult.RenderData) {
+      result->primitiveResults.pop_back();
+    }
   }
 }
 
@@ -1485,7 +1488,6 @@ static void loadPrimitiveGameThreadPart(
     UCesiumGltfComponent* pGltf,
     LoadPrimitiveResult& loadResult,
     const glm::dmat4x4& cesiumToUnrealTransform) {
-
   FName meshName = createSafeName(loadResult.name, "");
   UCesiumGltfPrimitiveComponent* pMesh =
       NewObject<UCesiumGltfPrimitiveComponent>(pGltf, meshName);